experiment(edge-preview): add bounded idle preload and four-body cross-view reuse [debug-ci]
Adopt the source and checks validated in Windows run 34640942241, plus disable keyboard navigation inside the hidden zero-size preload holder. Preserve the current renderer, excerpt limit, exact layout/resource keys, exclusive body ownership, cancellation and normal miss path.

Same-machine cold/text/layout comparisons and retained managed heap measurements are documented in tests/PaperTodo.EdgePreviewChecks/PRELOAD.md. All 24 cached/cold pixel fixtures match byte-for-byte; seven Release groups and Debug checks passed on the measured candidate. Preserve unfavorable text-only outcomes and the up-front preload costs.

Delete temporary assembly inputs and workflow. Keep this separate from #245/main and #238; final normal CI validates the runtime tree. No new backend, extra HWND, global mouse polling, background WPF thread or persistence protocol changes.
