--- .github/workflows/preview-preload-experiment.yml
@@ -1,80 +1,0 @@
-name: Preview preload experiment
-on:
-  push:
-    branches: [experiment/edge-preview-preload]
-    paths:
-      - .github/workflows/preview-preload-experiment.yml
-      - tests/preload-candidate/**
-permissions:
-  contents: write
-jobs:
-  validate:
-    runs-on: windows-latest
-    timeout-minutes: 20
-    defaults:
-      run:
-        shell: pwsh
-    steps:
-      - uses: actions/checkout@v7
-        with:
-          fetch-depth: 0
-          submodules: recursive
-      - uses: actions/setup-dotnet@v6
-        with:
-          dotnet-version: '10.0.x'
-      - name: Assemble isolated candidate
-        run: |
-          New-Item -ItemType Directory -Force results | Out-Null
-          python -X utf8 tests/preload-candidate/apply.py
-          if ($LASTEXITCODE -ne 0) { throw 'Assembly failed' }
-          python -X utf8 tests/preload-candidate/finish.py
-          if ($LASTEXITCODE -ne 0) { throw 'Cache validation assembly failed' }
-          git diff --check
-          if ($LASTEXITCODE -ne 0) { throw 'Whitespace failure' }
-      - name: Release behavior checks
-        run: |
-          $PSNativeCommandUseErrorActionPreference = $false
-          $results = @{}
-          foreach ($name in @('EdgePreviewChecks','MarkdownSemanticChecks','MarkdownEditingChecks','TodoNavigationChecks','EdgeTitleChecks','ThreadingChecks','PersistenceChecks')) {
-            dotnet run --project "tests/PaperTodo.$name" -c Release 2>&1 | Tee-Object "results/$name.log"
-            $results[$name] = $LASTEXITCODE
-          }
-          $results | ConvertTo-Json | Set-Content results/exit-codes.json
-          if (@($results.Values | Where-Object { $_ -ne 0 }).Count) { throw 'Behavior failures' }
-      - name: Cold, text-only and completed-layout forward comparison
-        run: |
-          dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release --no-build -- --preload-profile 2>&1 | Tee-Object results/profile-forward.log
-          if ($LASTEXITCODE -ne 0) { throw 'Forward profile failed' }
-      - name: Reverse-order comparison and retained memory
-        run: |
-          dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release --no-build -- --preload-profile --reverse 2>&1 | Tee-Object results/profile-reverse.log
-          if ($LASTEXITCODE -ne 0) { throw 'Reverse profile failed' }
-          dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release --no-build -- --preload-memory 2>&1 | Tee-Object results/memory.log
-          if ($LASTEXITCODE -ne 0) { throw 'Memory check failed' }
-      - name: Debug preload checks
-        run: |
-          dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Debug 2>&1 | Tee-Object results/debug.log
-          if ($LASTEXITCODE -ne 0) { throw 'Debug checks failed' }
-      - name: Record checked source without moving any branch
-        env:
-          GH_TOKEN: ${{ github.token }}
-        run: |
-          $headers = @{Authorization="Bearer $env:GH_TOKEN"; Accept='application/vnd.github+json'}
-          $paths = @((git diff --name-only)) + @('src/EdgeCapsulePreview.Preload.cs', 'src/PaperWindow.EdgePreviewPreload.cs', 'tests/PaperTodo.EdgePreviewChecks/PreloadChecks.cs')
-          $entries = foreach ($path in ($paths | Select-Object -Unique)) {
-            $text = [IO.File]::ReadAllText("$pwd/$path").Replace("`r`n","`n")
-            $body = @{content=$text; encoding='utf-8'} | ConvertTo-Json -Compress
-            $blob = Invoke-RestMethod -Method Post -Uri "https://api.github.com/repos/$env:GITHUB_REPOSITORY/git/blobs" -Headers $headers -Body $body -ContentType 'application/json; charset=utf-8'
-            @{path=$path; sha=$blob.sha; mode='100644'; type='blob'}
-            $destination = Join-Path "$pwd/results/source" $path
-            New-Item -ItemType Directory -Force (Split-Path $destination) | Out-Null
-            [IO.File]::WriteAllText($destination,$text,[Text.UTF8Encoding]::new($false))
-          }
-          $entries | ConvertTo-Json | Set-Content results/checked-blobs.json
-          git diff | Set-Content results/candidate.diff
-      - uses: actions/upload-artifact@v7
-        if: always()
-        with:
-          name: preview-preload-evidence
-          path: results/
-          retention-days: 2
