+++ src/PaperWindow.EdgeCapsulePreview.cs
@@ -55,6 +55,7 @@ public sealed partial class PaperWindow
         {
             return null;
         }
+        MarkdownEdgePreviewPreload.For(Dispatcher).BeginDemand();
         var prepareStartedAt = EdgeCapsulePerformanceDiagnostics.Timestamp();
         var bodySessionGeneration = _bodySessionGeneration;
 
@@ -130,6 +131,16 @@ public sealed partial class PaperWindow
                 return null;
             }
 
+            if (content is MarkdownEdgeCapsulePreviewView markdownView)
+            {
+                void Prepared(bool success)
+                {
+                    markdownView.PreloadViewport.PreparationFinished -= Prepared;
+                    if (success && _windowLifecycle == PaperWindowLifecycleState.Alive)
+                        _controller.ScheduleMarkdownPreviewNeighbors(this);
+                }
+                markdownView.PreloadViewport.PreparationFinished += Prepared;
+            }
             content.HorizontalAlignment = HorizontalAlignment.Stretch;
             content.VerticalAlignment = VerticalAlignment.Stretch;
             EdgeCapsulePerformanceDiagnostics.Trace(
