+++ src/PaperWindow.EdgePreviewPreload.cs
