+++ tests/PaperTodo.EdgePreviewChecks/PRELOAD.md
@@ -1,0 +1,65 @@
+# 有界预载实验
+
+本分支基于 #245 `ae1eb739464dc04e7cc23b8d749cb4ed7aa24f60`。预载是现有预览的可丢弃优化，不是另一套 Markdown 后端，也不改变 #238 的动画层。本实验不修改 main/#245 分支，不据此批准合并或正式发布。正式架构仍以 [ARCHITECTURE](../../doc/ARCHITECTURE.md) 为准；实验选型未确认，不把暂定缓存策略写成已接受的全局架构或正式发布承诺。
+
+## 运行行为
+
+原侧边悬停预览开启时，对存活、可见的内置笔记在变更后合并请求，180ms 单次延后后于 UI 空闲优先级准备。纯数据摘要最多 128 份，每份仍受 16 块/6000 字符限制；不是无限给所有笔记常驻控件。
+
+完整排版只保留最近 4 份未挂载的正文树，当前视图独占正在显示的正文。编辑/初始化产生的请求和已有队列顺序中的当前、前后相邻胶囊可请求预排版；正式悬停开始立即取消投机任务，不等待它完成。当前正文发布后才再请求邻居。预载未完成、被淘汰或条件不符时，继续 #245 的分批准备和一次发布，不显示半成品。
+
+预排版使用真实目标宿主内零尺寸、裁剪、透明、不可交互的临时容器继承 DPI/资源，不建立第二个窗口，不 StagePreview、不调整 HWND、不接管队列/命中/动画状态。最终版本还禁止这个隐藏容器参与 Tab、Ctrl+Tab 和方向键导航。容器在完成、取消、失效、隐藏和卸载路径移除。低优先级不等于后台线程，单个 WPF 工作段仍不可抢占。
+
+缓存键包含源身份、版本、实际摘要、模式、缩放、精确视口尺寸、DPI、字体/语言及资源颜色。取摘要时重新比较有界源内容，避免漏通知时沿用旧正文；描述与首次挂载间发生版本变化也会重新取内容。复杂/动画画刷不满足可复用条件时不命中。一个正文只能在一个视觉树中，不能被两个预览同时借用。
+
+## 已执行验证
+
+[运行 34640942241](https://github.com/snownico0722/PaperTodo/actions/runs/34640942241)，artifact `preview-preload-evidence`：七组 Release（EdgePreview、MarkdownSemantic、MarkdownEditing、TodoNavigation、EdgeTitle、Threading、Persistence）全部退出 0，Debug 预载检查通过。
+
+- A→B→A 三次正式打开均命中此前独立准备的完整正文；关闭后的重新建视图也可复用，不只同一 View 短暂恢复。
+- 24 组冷/热缓存像素对照字节完全相同，覆盖四档模式、两档缩放和三种短文/密集长文/块级输入；每组先断言实际命中缓存，不能用两次冷构建冒充。
+- 实际文字变化（含漏通知）、版本、主题颜色、尺寸、缩放、模式均拒绝旧结果；DPI 键失配拒绝，#245 的 DPI 通知检查仍运行。
+- 取消、过期结果、排队请求、4 份正文/128 份摘要上限及 Clear/移除临时容器检查通过。
+- 性能及像素数据采用该运行受测源码；最终提交只另外为隐藏预载容器增加键盘导航禁用，未修改排版算法和比较策略，随最终标准 CI 复验。没有宣称这些像素检查覆盖所有字体、真实多显示器或辅助功能。
+
+最初两轮临时拼装分别因 Windows 文本编码和不唯一替换定位失败，未进入运行代码验证；修正后第三、四轮运行完成。失败记录仍保留。一次性拼装文件/工作流已删除，维护测试保留在现有 EdgePreviewChecks 中。
+
+## 同机三种策略对照
+
+同一 Windows runner、相同二进制和真实 Host/Presenter、460×410 卡片、160ms 动画。每种内容/模式分别运行不预载、只预解析、完整排版预载；第二个进程反转策略顺序。每次策略 9 个新视图样本，排除前 2 个，每种策略每种场景共 14 个正式样本。不是 14 次应用冷启动，也不是可靠线上 p95。
+
+下表是开始计算卡片尺寸到正文布局就绪的中位数（ms）。完整排版列只测已预载成功的命中路径，全部 84 个正式命中样本均记录到 1 次命中；这是测试条件，不是实际使用的 100% 命中率。
+
+| 场景 | 模式 | 不预载 | 仅预解析 | 完整排版命中 | 提前完成排版耗时 |
+|---|---|---:|---:|---:|---:|
+| 普通短行 | 增强 | 22.21 | 22.03 | 12.94 | 24.27 |
+| 密集长行 | 增强 | 72.43 | 65.96 | 1.63 | 72.98 |
+| 短密集多行 | 增强 | 30.46 | 32.37 | 0.98 | 31.95 |
+| 普通短行 | 完全渲染 | 10.55 | 10.76 | 7.85 | 14.33 |
+| 密集长行 | 完全渲染 | 56.97 | 58.93 | 1.05 | 49.29 |
+| 短密集多行 | 完全渲染 | 25.88 | 27.38 | 0.89 | 19.24 |
+
+预解析不是全项更快，密集长行的增强模式略有改善，完全渲染和短密集多行有更慢样本。最明显的收益来自完整排版结果的复用，不是继续优化语法解析。
+
+预排版不是免费：表中最后一列是打开前已付出的等待成本（包含协作等待），不能加减两列中位数作为逐样本总 CPU 时间。猜错目标或失效会浪费工作；首次启动立即悬停、迅速滑过或修改后立刻打开仍可能不命中。
+
+完整排版命中后的最大应用侧动画回调间隔中位数约 13.93–15.81ms；不预载约 17.63–22.92ms。这不是 GPU 最终呈现、显示器帧率或 240Hz 丝滑证明，也未穷尽预载时对别的活动窗口动画的影响。普通短行的完整控件重挂载仍有布局成本，不承诺所有命中都是 1ms。
+
+前一轮 [34640529500](https://github.com/snownico0722/PaperTodo/actions/runs/34640529500) 在不同 runner 上完整渲染密集长行 88.08→1.36ms，趋势相同，但不把两台 runner 的绝对值拼接成一张改进表。
+
+## 实际内存测量
+
+在样本字符串/上下文已创建并保留、解析器预热后，比较完整 GC 后的托管存活堆：100 份各 6000 字符且包含密集样式的摘要增加 **6384.01KiB（6.23MiB）**；再在真实窗口下保留 4 份排版，连同该阶段 WPF 内部缓存增加 **4739.53KiB（4.63MiB）**。两项合计约 **10.86MiB**。
+
+源样本文本已在测量起点存活，故不重复计入。后一个值不是精确的四棵正文树大小；这不是进程私有内存、原生字形/合成器/GPU 资源或工作集，也不是所有最坏输入的内存上界。容量限制是摘要 128 / 脱离视图的正文 4；当前显示和一个正在准备的对象另有短期占用。
+
+## 复跑
+
+```powershell
+dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release
+dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release -- --preload-profile
+dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release -- --preload-profile --reverse
+dotnet run --project tests/PaperTodo.EdgePreviewChecks -c Release -- --preload-memory
+```
+
+`PRELOAD_SAMPLE.metrics` 顺序是 describe、create、stage、totalReady、stageToReady、firstShape、frameGapMax、applyMax、demandAllocationKiB、warmElapsed、bodyHits。分配量只覆盖正式打开阶段，不是预载加打开的总分配；原始 warm 值、反例和全部样本均保留。最终是否采用仍需真实浏览命中率、快速跨卡片操作和混合 DPI 体验验证。
