+++ tests/PaperTodo.EdgePreviewChecks/PreloadChecks.cs
@@ -3,6 +3,7 @@ using System.Text.Json;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
+using System.Windows.Media.Imaging;
 using System.Windows.Threading;
 using PaperTodo;
 
@@ -61,6 +62,63 @@ internal static partial class Program
             Release(Demand(a)); Release(Demand(b)); Release(Demand(a));
             Require(cache.BodyHits == hits + 3, "A-B-A uses independently owned completed bodies three times");
             Console.WriteLine("PASS preload A-B-A and detached return: three demand hits");
+            var pixelFixtures = new[] {
+                "plain **strong** *italic* `code` [link](https://example.com)",
+                string.Concat(Enumerable.Repeat("**粗体** ~~删除~~ `code` [链接](https://example.com) 文 ", 80)),
+                "## heading\n- [x] **done**\n> quote *text*\n```\n\ncode\n```"
+            };
+            var pixelCases = 0;
+            var exactCases = 0;
+            var maximumDifference = 0;
+            byte[] Pixels(FrameworkElement element)
+            {
+                var dpi = VisualTreeHelper.GetDpi(element);
+                int width = (int)Math.Ceiling(element.ActualWidth * dpi.DpiScaleX);
+                int height = (int)Math.Ceiling(element.ActualHeight * dpi.DpiScaleY);
+                var bitmap = new RenderTargetBitmap(width, height, 96 * dpi.DpiScaleX, 96 * dpi.DpiScaleY, PixelFormats.Pbgra32);
+                bitmap.Render(element);
+                var bytes = new byte[width * height * 4];
+                bitmap.CopyPixels(bytes, width * 4, 0);
+                return bytes;
+            }
+            foreach (var testMode in new[] { MarkdownRenderModes.Off, MarkdownRenderModes.Basic, MarkdownRenderModes.Enhanced, MarkdownRenderModes.Full })
+            foreach (var zoom in new[] { 0.7, 1.3 })
+            foreach (var fixture in pixelFixtures)
+            {
+                mode = testMode; text = fixture; cache.Clear();
+                var context = Context(new()); context.Paper.TextZoom = zoom;
+                Require(Warm(context), "pixel reference is actually preloaded");
+                var previousHits = cache.BodyHits;
+                var hot = Demand(context); var hotPixels = Pixels(hot); Release(hot);
+                Require(cache.BodyHits == previousHits + 1, "pixel comparison traverses the cached-body path");
+                cache.Clear();
+                var cold = Demand(context); var coldPixels = Pixels(cold); Release(cold);
+                Require(hotPixels.Length == coldPixels.Length, "cache preserves pixel dimensions");
+                var differences = hotPixels.Zip(coldPixels).Select(pair => Math.Abs(pair.First - pair.Second)).ToArray();
+                maximumDifference = Math.Max(maximumDifference, differences.Max());
+                if (hotPixels.SequenceEqual(coldPixels)) exactCases++;
+                Require(differences.All(delta => delta <= 32), "cache preserves visible text/decoration pixels");
+                pixelCases++;
+            }
+            Console.WriteLine($"PRELOAD_PIXELS cases={pixelCases} exact={exactCases} maximumChannelDifference={maximumDifference}");
+            mode = MarkdownRenderModes.Full;
+            cache.Clear();
+            var descriptorBeforeEdit = MarkdownEdgeCapsulePreviewProvider.Instance.Describe(a);
+            text = "edited between describe and mount";
+            source.Invalidate();
+            var editedView = descriptorBeforeEdit.CreateContent(size);
+            var editedBorder = new Border { Width = size.WidthDip - 22, Height = size.HeightDip, Child = editedView };
+            root.Children.Add(editedBorder);
+            ((EdgeCapsuleLivePreviewView)editedView).PrepareForFirstDisplay();
+            Pump();
+            Require(PreviewText(editedView).Contains(text), "deferred first display never binds an old excerpt to the new version");
+            Release(editedBorder);
+            var binding = cache.Bind(a, cache.Capture(a), a.Paper.TextZoom)!;
+            var key = MarkdownEdgePreviewPreload.MakeKey(binding, root, new Size(200, 100))!;
+            Require(cache.Store(new(key, new StackPanel(), false)), "cache key fixture retained");
+            Require(!cache.TryTake(key with { Dpi = new DpiScale(key.Dpi.DpiScaleX * 1.5, key.Dpi.DpiScaleY * 1.5) }, out _), "different DPI cannot reuse old drawing");
+            Require(cache.TryTake(key, out _, demand: false), "matching DPI key remains usable");
+            Console.WriteLine("PASS first-display generation and DPI-key rejection");
             hits = cache.BodyHits;
             text = "新内容，不允许用旧正文";
             var changed = Demand(a);
